refactor(#186): SIP-0097 slice 3 — correction protocol moves to CorrectionRunner
Extract _run_correction_protocol + _store_correction_task_artifacts +
_checkpoint_correction_task verbatim into adapters/cycles/correction_runner.py
(SIP-0097 §6.3). Outcome routing (_handle_task_outcome) stays with the
executor's orchestration loop.

Interim seam per §6.3/AC#9: the runner receives three narrow executor-supplied
callables — dispatch_task + create_task_run (re-pointed to TaskDispatcher in
slice 5) and store_artifact (§6.7 executor residual). Late-bound lambdas so
tests that stub _dispatch_task on the instance keep working. No cancellation
probe: the protocol never checked cancellation; it relies on dispatch-boundary
checks per the §6 cancellation ownership rule.

Mechanical test updates (disclosed per §7.7): the two correction test
fixtures now pass event_bus at construction (the composed default runner
captures it in __init__; post-hoc assignment no longer reaches correction
events); direct protocol/helper call sites re-pointed to the collaborator.
Emission guardrails re-anchored: correction_runner.py joins the emission
file set; executor 20→17 EventType refs; total emit sites unchanged at 44
(27 executor + 10 correction-runner + 7 route).

run_correction_protocol passes C901=12 at its new home — no new per-file
ignore; pyproject.toml untouched. Executor: 3,038 → 2,632 lines.
